requirements.txt
django==1.6.7
evernote
markdown
pillow
south
django-tastypie